added collision

+++ src/crepe/api/CircleCollider.h
@@ -5,9 +5,9 @@ namespace crepe::api {
 
 class CircleCollider : public Collider {
 public:
-	CircleCollider(uint32_t game_object_id, int radius)
+	CircleCollider(uint32_t game_object_id, double radius)
 		: Collider(game_object_id), radius(radius) {}
-	int radius;
+	double radius;
 };
 
 } // namespace crepe::api
